데이터 커밋
데이터 커밋

+++ 객관식_만들기.py
@@ -4,33 +4,50 @@ import googletrans
 from tqdm import tqdm
 import numpy as np
 import os
+from os.path import exists, isfile, basename
 
-
-def 엑셀파일구분하기(data_direct, filename):
+def 엑셀파일구분하기(data_direct, filename,보존 = False,구분_나누기=True,날짜_나누기=True):
     data = pd.read_excel(f"{data_direct}{filename}.xlsx")
     return_list1 = []
     return_list2 = []
-    try:
-        날짜목록 = list(set(data["날짜"]))
-        for 날짜 in 날짜목록:
-            save_filename1 = f"{data_direct}{filename}_{날짜}.xlsx"
-            data1 = data[data["날짜"] == 날짜]
-            data1.to_excel(save_filename1, index=False)
+    if 날짜_나누기 == True:
+        try:
+            날짜목록 = list(set(data["날짜"]))
+            for 날짜 in 날짜목록:
+                save_filename1 = f"{data_direct}{filename}_{날짜}.xlsx"
+                data1 = data[data["날짜"] == 날짜]
+                if 보존 == True:
+                    if isfile(save_filename1):
+                        pass
+                    else:
+                        data1.to_excel(save_filename1, index=False)
+                        print(save_filename1)
+                else:
+                    data1.to_excel(save_filename1, index=False)
+                    print(save_filename1)
             return_list1 += [f"{filename}_{날짜}"]
-            print(save_filename1)
-    except:
-        print("날짜 없음!")
+        except:
+            print("날짜 없음!")
 
-    try:
-        구분목록 = list(set(data["구분"]))
-        for 구분 in 구분목록:
-            save_filename2 = f"{data_direct}{filename}_{구분}.xlsx"
-            data2 = data[data["구분"] == 구분]
-            data2.to_excel(save_filename2, index=False)
-            return_list2 += [f"{filename}_{구분}"]
-            print(save_filename2)
-    except:
-        print("구분 없음!")
+    if 구분_나누기 == True:
+        try:
+            구분목록 = list(set(data["구분"]))
+            for 구분 in 구분목록:
+                save_filename2 = f"{data_direct}{filename}_{구분}.xlsx"
+                data2 = data[data["구분"] == 구분]
+                if 보존 == True:
+                    if isfile(save_filename2):
+                        pass
+                    else:
+                        data2.to_excel(save_filename2, index=False)
+                        print(save_filename2)
+                else:
+                    data2.to_excel(save_filename2, index=False)
+                    print(save_filename2)
+                return_list2 += [f"{filename}_{구분}"]
+                print(save_filename2)
+        except:
+            print("구분 없음!")
     return return_list1, return_list2
 
 
@@ -88,6 +105,27 @@ def 객관식_만들기_한자어(파일명, data_direct, 단답형=True, 설명
     print(save_filename)
     save_data.to_excel(save_filename, index=False)
 
+def 단답형_만들기_한자어(파일명, data_direct, 단답형=True, 설명=True, 글자수=2, 번역=False):
+    translator = googletrans.Translator()
+    data = pd.read_excel(data_direct+파일명+".xlsx")
+    한자 = []
+    한글 = []
+    for i in tqdm(range(data["대답"].size)):
+        질문 = data["질문"][i]
+        대답 = data["대답"][i]
+        if len(질문.split('/')[0]) == 글자수:
+            if 번역 == True:
+                한자번역 = translator.translate(질문, dest='en')
+                한자번역 = 한자번역.text
+                if len(한자번역.split(" ")) == 1 and 한자번역 != "no":
+                    대답 = f"{대답}({한자번역})"
+            한자 += [질문]
+            한글 += [대답]
+    data = pd.DataFrame({"질문": 한자, "대답": 한글})
+
+    save_filename = data_direct +  파일명 + f"{글자수}글자{번역}.xlsx"
+    print(save_filename)
+    data.to_excel(save_filename, index=False)
 
 def 객관식_만들기(파일명, data_direct, 단답형=True, 설명=True):
     data = pd.read_excel(data_direct+파일명+".xlsx")
@@ -330,16 +368,65 @@ def OX퀴즈만들기(data_direct, filename):
     data.to_excel(
         f"{data_direct}{filename}.xlsx".replace("단답형", "O X 퀴즈"))
 
+#############################################
+# 초성 리스트. 00 ~ 18
+CHOSUNG_LIST = ['ㄱ', 'ㄲ', 'ㄴ', 'ㄷ', 'ㄸ', 'ㄹ', 'ㅁ', 'ㅂ', 'ㅃ', 'ㅅ', 'ㅆ', 'ㅇ', 'ㅈ', 'ㅉ', 'ㅊ', 'ㅋ', 'ㅌ', 'ㅍ', 'ㅎ']
+# 중성 리스트. 00 ~ 20
+JUNGSUNG_LIST = ['ㅏ', 'ㅐ', 'ㅑ', 'ㅒ', 'ㅓ', 'ㅔ', 'ㅕ', 'ㅖ', 'ㅗ', 'ㅘ', 'ㅙ', 'ㅚ', 'ㅛ', 'ㅜ', 'ㅝ', 'ㅞ', 'ㅟ', 'ㅠ', 'ㅡ', 'ㅢ',
+                 'ㅣ']
+# 종성 리스트. 00 ~ 27 + 1(1개 없음)
+JONGSUNG_LIST = [' ', 'ㄱ', 'ㄲ', 'ㄳ', 'ㄴ', 'ㄵ', 'ㄶ', 'ㄷ', 'ㄹ', 'ㄺ', 'ㄻ', 'ㄼ', 'ㄽ', 'ㄾ', 'ㄿ', 'ㅀ', 'ㅁ', 'ㅂ', 'ㅄ', 'ㅅ',
+                 'ㅆ', 'ㅇ', 'ㅈ', 'ㅊ', 'ㅋ', 'ㅌ', 'ㅍ', 'ㅎ']
 
+
+def korean_to_be_englished(korean_word):
+    r_lst = []
+    for w in list(korean_word.strip()):
+        ## 영어인 경우 구분해서 작성함.
+        if '가' <= w <= '힣':
+            ## 588개 마다 초성이 바뀜.
+            ch1 = (ord(w) - ord('가')) // 588
+            ## 중성은 총 28가지 종류
+            ch2 = ((ord(w) - ord('가')) - (588 * ch1)) // 28
+            ch3 = (ord(w) - ord('가')) - (588 * ch1) - 28 * ch2
+            r_lst.append([CHOSUNG_LIST[ch1], JUNGSUNG_LIST[ch2], JONGSUNG_LIST[ch3]])
+        else:
+            r_lst.append([w])
+    return r_lst
+
+
+#korean_to_be_englished("이승훈a")
+#############################################
+def 구분_생성(data_direct, filename):
+    df = pd.read_excel(f"{data_direct}{filename}.xlsx")
+    if '구분' not in list(df.keys()):
+        df['구분'] = ['기본' for i in range(len(df))]
+    df['구분'] = df['구분'].fillna('기본')
+    num_total = df.shape[0]
+    for i in range(num_total):
+        대답 = df['대답'][i]
+        대답 = 대답.rstrip()
+        '''
+        if korean_to_be_englished(대답)[-1][-1] != ' ':
+            df['구분'][i] = korean_to_be_englished(대답)[-1][-1]
+        else:
+            df['구분'][i] = korean_to_be_englished(대답)[-2][-1]
+        '''
+        df['구분'][i] = 대답[-1]
+    df.to_excel(f"{data_direct}{filename}.xlsx",index=False)
 data_direct = "./학습자료/단답형/"
-filename = "조선"
+filename = "국어_암기자료"
 if filename == "국어_복습":
     엑셀파일구분하기(data_direct, filename)
+    단답형_만들기_한자어("국어_복습_한자어", data_direct, 단답형=False,설명=False,글자수=1,번역= False)
+    단답형_만들기_한자어("국어_복습_한자어", data_direct, 단답형=False,설명=False,글자수=2,번역=True)
+    단답형_만들기_한자어("국어_복습_한자어", data_direct, 단답형=False,설명=False,글자수=3,번역=True)
+    단답형_만들기_한자어("국어_복습_한자어", data_direct, 단답형=False,설명=False,글자수=4,번역=False)
+    '''
     객관식_만들기("국어_복습_속담", data_direct, 단답형=False)
     객관식_만들기("국어_복습_의미", data_direct, 단답형=False, 설명=False)
     특정구분제거(data_direct, filename, 구분목록=["속담"])
     객관식_만들기_한자어("국어_복습_한자어", data_direct, 단답형=False, 설명=False, 글자수=4, 번역=False)
-    '''
     OX퀴즈만들기(data_direct, "국어_복습_한자어")
     객관식_만들기_한자어("국어_복습_한자어", data_direct, 단답형=False,설명=False,글자수=1,번역=False)
     객관식_만들기_한자어("국어_복습_한자어", data_direct, 단답형=False,설명=False,글자수=2,번역=True)
@@ -347,12 +434,14 @@ if filename == "국어_복습":
     객관식_만들기_한자어("국어_복습_한자어", data_direct, 단답형=False,설명=False,글자수=4,번역=False)
     '''
 elif filename == "국어_암기자료":
-    객관식_만들기_구분통합(filename, data_direct, 단답형=False, 설명=True)
+    엑셀파일구분하기(data_direct, filename,보존=True,구분_나누기=False,날짜_나누기=True)
 elif filename == "국어_57항":
     객관식_만들기_구분통합(filename, data_direct, 단답형=False, 설명=True)
 elif filename == "영어_단어":
-    객관식_만들기_구분통합(filename, data_direct, 단답형=False, 설명=True)
+    #구분_생성(data_direct, filename)
+    엑셀파일구분하기(data_direct, filename,보존=True,구분_나누기=False,날짜_나누기=True)
     '''
+    객관식_만들기_구분통합(filename, data_direct, 단답형=False, 설명=True)
     return_list1, return_list2 = 엑셀파일구분하기(data_direct, filename)
     객관식_만들기(filename, data_direct, 단답형=False,설명=True)
     for filename in return_list1:
